feat: AddPlay UI Composable Screen (#98)
* feat: add AddPlayScreen composable with game search and full play form

Agent-Logs-Url: https://github.com/trumpets/meeplebook/sessions/1273a0f1-9415-4d98-841b-633f0e562db6


* wired in navigation contract for add play

* changes to UI to have location suggestions as dropdown and moved player chips around

* dialog before discarding play.

* trying some location focus things. don't work. will circle back

* date + duration in same row

* added quantity, incomplete and comments

* reworked player rows to be fully functional. dragging has some quirks. gonna solve it.

* feat: score as Double, usedColorsForGame in state, ObserveColorsUsedForGameUseCase, update tests

Agent-Logs-Url: https://github.com/trumpets/meeplebook/sessions/bb716fa9-4caf-4c28-a55d-7ccad1ae4df2


* playerscore is double

* refactored VM to structure it better. it's a large one

* some renaming

* preselectedgame for addplay screen

* preselectedgame had a serialization problem in nav. decomposed it while navigating

* some fixes

* color picker dialog and score input dialog

* some fixes

* split screen in multiple files

* Update app/src/main/java/app/meeplebook/core/plays/model/PlayerColor.kt


* previews for all

* AddPlaySections grouping Kotlin hack

* all tests pass now

* Update app/src/main/java/app/meeplebook/feature/addplay/AddPlayScreen.kt


* preview fix

* fix britain!

* millis fix

---------

+++ app/src/androidTest/java/app/meeplebook/core/database/dao/PlayerDaoTest.kt
@@ -651,7 +651,7 @@ class PlayerDaoTest {
         userId: Long? = null,
         startPosition: Int? = null,
         color: String? = null,
-        score: Int? = null
+        score: Double? = null
     ): PlayerEntity {
         return PlayerEntity(
             id = id,
