+++ app/src/main/java/app/meeplebook/core/database/entity/PlayerEntity.kt
@@ -36,7 +36,7 @@ data class PlayerEntity(
     val name: String,
     val startPosition: Int?,
     val color: String?,
-    val score: Int?,
+    val score: Double?,
     val win: Boolean
 )
